+++ tests/unit_tests/test_config.py
@@ -4,6 +4,7 @@ import logging
 
 import pytest
 import yaml
+
 from casanovo.config import Config